+++ src/main/java/mekanism/common/integration/tesla/TeslaIntegration.java
@@ -32,14 +32,14 @@ public class TeslaIntegration implements ITeslaHolder, ITeslaConsumer, ITeslaPro
 	@Method(modid = MekanismHooks.TESLA_MOD_ID)
 	public long getStoredPower() 
 	{
-		return (long)Math.round(tileEntity.getEnergy()*general.TO_TESLA);
+		return Math.round(tileEntity.getEnergy()*general.TO_TESLA);
 	}
 
 	@Override
 	@Method(modid = MekanismHooks.TESLA_MOD_ID)
 	public long getCapacity() 
 	{
-		return (long)Math.round(tileEntity.getMaxEnergy()*general.TO_TESLA);
+		return Math.round(tileEntity.getMaxEnergy()*general.TO_TESLA);
 	}
 	
 	@Override
@@ -58,7 +58,7 @@ public class TeslaIntegration implements ITeslaHolder, ITeslaConsumer, ITeslaPro
 	
 	public long rfToTesla(int rf)
 	{
-		return (long)Math.round(rf*general.FROM_RF*general.TO_TESLA);
+		return Math.round(rf*general.FROM_RF*general.TO_TESLA);
 	}
 	
 	public int teslaToRF(long tesla)
